+++ lib/src/types/callbacks.dart
@@ -12,7 +12,6 @@ typedef DragEndCallback<T> = void Function(T mapObject);
 typedef ClusterCallback = Future<Cluster?> Function(ClusterizedPlacemarkCollection self, Cluster cluster);
 typedef ClusterTapCallback = void Function(ClusterizedPlacemarkCollection self, Cluster cluster);
 typedef MapCreatedCallback = void Function(YandexMapController controller);
-typedef MapRenderedCallback = void Function(YandexMapController controller, MapSize size);
 typedef SearchErrorCallback = void Function(String msg, int sessionId);
 typedef CancelDrivingSessionCallback = void Function();
 typedef UserLocationCallback = Future<UserLocationView>? Function(UserLocationView view);